revamp install script

+++ install.sh
@@ -18,6 +18,8 @@ Dim='' # White
 Bold_White=''
 Bold_Green=''
 
+echo "$@"
+
 if [[ -t 1 ]]; then
     # Reset
     Color_Off='\033[0m' # Text Reset
@@ -86,10 +88,10 @@ exe_name=duckdb
 duckdb_uri=$github_repo/$target.zip
 
 install_env=DUCKDB_INSTALL
-bin_env=\$$install_env/bin
+bin_env=\$$install_env/latest/bin
 
 install_dir=${!install_env:-$HOME/.duckdb}
-bin_dir=$install_dir/bin
+bin_dir=$install_dir/latest/bin
 exe=$bin_dir/duckdb-latest
 
 if [[ ! -d $bin_dir ]]; then
@@ -97,19 +99,25 @@ if [[ ! -d $bin_dir ]]; then
         error "Failed to create install directory \"$bin_dir\""
 fi
 
-curl --fail --location --progress-bar --output "$exe.zip" "$duckdb_uri" ||
-    error "Failed to download duckdb from \"$duckdb_uri\""
 
-unzip -oqd "$bin_dir" "$exe.zip" ||
-    error 'Failed to extract duckdb'
+if [[ $(find "$exe" -mtime +1 -print) ]]; then
+  curl --fail --location --progress-bar --output "$exe.zip" "$duckdb_uri" ||
+      error "Failed to download duckdb from \"$duckdb_uri\""
 
-mv "$bin_dir/$exe_name" "$exe" ||
-    error 'Failed to move extracted duckdb to destination'
+  unzip -oqd "$bin_dir" "$exe.zip" ||
+      error 'Failed to extract duckdb'
 
-chmod +x "$exe" ||
-    error 'Failed to set permissions on duckdb executable'
+  mv "$bin_dir/$exe_name" "$exe" ||
+      error 'Failed to move extracted duckdb to destination'
 
-rm "$exe.zip"
+  chmod +x "$exe" ||
+      error 'Failed to set permissions on duckdb executable'
+
+  rm "$exe.zip"
+  
+  touch "$exe"
+fi
+
 
 tildify() {
     if [[ $1 = $HOME/* ]]; then
@@ -122,3 +130,142 @@ tildify() {
 }
 
 success "duckdb was installed successfully to $Bold_Green$(tildify "$exe")"
+
+refresh_command=''
+
+tilde_bin_dir=$(tildify "$bin_dir")
+quoted_install_dir=\"${install_dir//\"/\\\"}\"
+
+if [[ $quoted_install_dir = \"$HOME/* ]]; then
+    quoted_install_dir=${quoted_install_dir/$HOME\//\$HOME/}
+fi
+
+echo
+
+case $(basename "$SHELL") in
+fish)
+    commands=(
+        "set --export $install_env $quoted_install_dir"
+        "set --export PATH $bin_env \$PATH"
+    )
+
+    fish_config=$HOME/.config/fish/config.fish
+    tilde_fish_config=$(tildify "$fish_config")
+
+    if [[ -w $fish_config ]]; then
+        {
+            echo -e '\n# bun'
+
+            for command in "${commands[@]}"; do
+                echo "$command"
+            done
+        } >>"$fish_config"
+
+        info "Added \"$tilde_bin_dir\" to \$PATH in \"$tilde_fish_config\""
+
+        refresh_command="source $tilde_fish_config"
+    else
+        echo "Manually add the directory to $tilde_fish_config (or similar):"
+
+        for command in "${commands[@]}"; do
+            info_bold "  $command"
+        done
+    fi
+    ;;
+zsh)
+    commands=(
+        "export $install_env=$quoted_install_dir"
+        "export PATH=\"$bin_env:\$PATH\""
+    )
+
+    zsh_config=$HOME/.zshrc
+    tilde_zsh_config=$(tildify "$zsh_config")
+
+    if [[ -w $zsh_config ]]; then
+        {
+            echo -e '\n# bun'
+
+            for command in "${commands[@]}"; do
+                echo "$command"
+            done
+        } >>"$zsh_config"
+
+        info "Added \"$tilde_bin_dir\" to \$PATH in \"$tilde_zsh_config\""
+
+        refresh_command="exec $SHELL"
+    else
+        echo "Manually add the directory to $tilde_zsh_config (or similar):"
+
+        for command in "${commands[@]}"; do
+            info_bold "  $command"
+        done
+    fi
+    ;;
+bash)
+    commands=(
+        "export $install_env=$quoted_install_dir"
+        "export PATH=$bin_env:\$PATH"
+    )
+
+    bash_configs=(
+        "$HOME/.bashrc"
+        "$HOME/.bash_profile"
+    )
+
+    if [[ ${XDG_CONFIG_HOME:-} ]]; then
+        bash_configs+=(
+            "$XDG_CONFIG_HOME/.bash_profile"
+            "$XDG_CONFIG_HOME/.bashrc"
+            "$XDG_CONFIG_HOME/bash_profile"
+            "$XDG_CONFIG_HOME/bashrc"
+        )
+    fi
+
+    set_manually=true
+    for bash_config in "${bash_configs[@]}"; do
+        tilde_bash_config=$(tildify "$bash_config")
+
+        if [[ -w $bash_config ]]; then
+            {
+                echo -e '\n# bun'
+
+                for command in "${commands[@]}"; do
+                    echo "$command"
+                done
+            } >>"$bash_config"
+
+            info "Added \"$tilde_bin_dir\" to \$PATH in \"$tilde_bash_config\""
+
+            refresh_command="source $bash_config"
+            set_manually=false
+            break
+        fi
+    done
+
+    if [[ $set_manually = true ]]; then
+        echo "Manually add the directory to $tilde_bash_config (or similar):"
+
+        for command in "${commands[@]}"; do
+            info_bold "  $command"
+        done
+    fi
+    ;;
+*)
+    echo 'Manually add the directory to ~/.bashrc (or similar):'
+    info_bold "  export $install_env=$quoted_install_dir"
+    info_bold "  export PATH=\"$bin_env:\$PATH\""
+    ;;
+esac
+
+"$exe" "$@"
+
+echo
+info "To get started, run:"
+echo
+
+if [[ $refresh_command ]]; then
+    info_bold " $refresh_command"
+fi
+
+info_bold "  bun --help"
+
